+++ src/lockfile.rs
@@ -19,8 +19,7 @@ impl Lockfile {
             grpc_port,
             http_port,
         };
-        let json = serde_json::to_string(&lock)
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))?;
+        let json = serde_json::to_string(&lock).map_err(std::io::Error::other)?;
         std::fs::write(lockfile_path(), json)
     }